+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -9,90 +9,37 @@ final class MovieQuizViewController: UIViewController {
     @IBOutlet private var yesButton: UIButton!
     @IBOutlet private var noButton: UIButton!
     
-    // MARK: - Structures
-    struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
-    }
-    
-    struct QuizStepViewModel {
-        let image: UIImage
-        let question: String
-        let questionNumber: String
-    }
-    
-    struct QuizResultsViewModel {
-        let title: String
-        let text: String
-        let buttonText: String
-    }
-    
     // MARK: - Private Properties
     private var currentQuestionIndex = 0
     private var correctAnswers = 0
+    private let questionsAmount = 10
+    private var questionFactory: QuestionFactory = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
+
     
-    // массив мок-данных
-    private var  questions: [QuizQuestion] = [
-        QuizQuestion(
-            image:"The Godfather",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Dark Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image:  "Kill Bill",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Avengers",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Deadpool",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Green Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Old",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "The Ice Age Adventures of Buck Wild",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Tesla",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Vivarium",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-    ]
     
     // MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
         imageView.layer.cornerRadius = 20
-        show(quiz: convert(model: questions[currentQuestionIndex]))
+        if let firstQuestion = questionFactory.requestNextQuestion() {
+            currentQuestion = firstQuestion
+            show(quiz: convert(model: firstQuestion))
+        }
     }
     
     // MARK: - IBActions
     @IBAction private func yesButtonClicked(_ sender: Any) {
+        guard let currentQuestion = currentQuestion else { return }
+      //отладочное  print(currentQuestion.correctAnswer, currentQuestion.image.description)
         let givenAnswer = true
-        let currentQuestion = questions[currentQuestionIndex]
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
     }
     
     @IBAction private func noButtonClicked(_ sender: Any) {
+        guard let currentQuestion = currentQuestion else { return }
+     //отладочное   print(currentQuestion.correctAnswer, currentQuestion.image.description)
         let givenAnswer = false
-        let currentQuestion = questions[currentQuestionIndex]
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
     }
     
@@ -102,7 +49,8 @@ final class MovieQuizViewController: UIViewController {
         let questionStep = QuizStepViewModel(
             image: UIImage(named: model.image) ?? UIImage(),
             question: model.text,
-            questionNumber: ("\(currentQuestionIndex + 1)/\(questions.count)"))
+            questionNumber: ("\(currentQuestionIndex + 1)/\(questionsAmount)"))
+        
         return questionStep
     }
     
@@ -124,22 +72,26 @@ final class MovieQuizViewController: UIViewController {
             guard let self = self else { return }
             self.currentQuestionIndex = 0
             self.correctAnswers = 0
-            
-            self.show(quiz: self.convert(model: self.questions[self.currentQuestionIndex]))
+            guard let currentQuestion = currentQuestion else { return }
+            show(quiz: convert(model: currentQuestion))
         }
         alert.addAction(action)
         self.present(alert, animated: true)
     }
     
     private func showNextQuestionOrResults() {
-        if currentQuestionIndex == questions.count - 1 {
+        guard let nextQuestion = questionFactory.requestNextQuestion() else { return }
+        currentQuestion = nextQuestion
+        if currentQuestionIndex == questionsAmount - 1 {
             show(quiz: .init(
                 title:  "Этот раунд окончен!",
-                text: "Ваш результат: \(correctAnswers)/10",
+                text: correctAnswers == questionsAmount ?
+                "Поздравляем, вы ответили на \(questionsAmount) из \(questionsAmount)!" :
+                "Вы ответили на \(correctAnswers) из \(questionsAmount), попробуйте ещё раз!",
                 buttonText: "Сыграть ещё раз"))
         } else {
             currentQuestionIndex += 1
-            show(quiz: convert(model: questions[currentQuestionIndex]))
+            show(quiz: convert(model: nextQuestion))
         }
     }
     
@@ -151,7 +103,7 @@ final class MovieQuizViewController: UIViewController {
         if isCorrect { correctAnswers += 1 }
         DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) { [weak self] in
             guard let self = self else { return }
-            self.showNextQuestionOrResults()
+            showNextQuestionOrResults()
         }
     }
